@@ -1,13 +1,14 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.1.3'
+gem 'rails', '3.2.6'
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails', ">= 3.1.4"
-  gem 'coffee-rails', "~> 3.1.0"
-  gem 'uglifier'
+  gem 'sass-rails', '~> 3.2.3'
+  gem 'coffee-rails', '~> 3.2.1'
+  gem 'therubyracer'
+  gem 'uglifier', '>= 1.0.3'
 end
 
 # Bundle edge Rails instead:
@@ -35,14 +36,16 @@ end
 # group :development, :test do
 #   gem 'webrat'
 # end
-gem "mongoid", "~> 2.3"
-gem "bson_ext", "~> 1.4"
-gem "jquery-rails"
-gem "omniauth", "~> 0.2.0"
-gem "hubruby"
-gem "faraday"
-gem "thin"
+
+gem 'mongoid'
+gem 'bson_ext'
+gem 'jquery-rails'
+gem 'omniauth'
+gem 'omniauth-github'
+gem 'hubruby'
+gem 'faraday'
+gem 'thin'
 
 group :development, :test do
-  gem "mocha"
+  gem 'mocha'
 end
